pyproject.toml
[project]
name = "dados"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "firebase-admin>=7.1.0",
    "pandas>=2.3.3",
    "plotly-express>=0.4.1",
    "streamlit>=1.50.0",
    "watchdog>=6.0.0",
]

[tool.uv]
package-mode = false
